feat (gui): starting screen returns graph to main screen
